src/leetcode/Solution0015.java
package leetcode;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

/**
 * 示例 1：
 *
 * 输入：nums = [-1,0,1,2,-1,-4]
 * 输出：[[-1,-1,2],[-1,0,1]]
 * 解释：
 * nums[0] + nums[1] + nums[2] = (-1) + 0 + 1 = 0 。
 * nums[1] + nums[2] + nums[4] = 0 + 1 + (-1) = 0 。
 * nums[0] + nums[3] + nums[4] = (-1) + 2 + (-1) = 0 。
 * 不同的三元组是 [-1,0,1] 和 [-1,-1,2] 。
 * 注意，输出的顺序和三元组的顺序并不重要。
 * 示例 2：
 *
 * 输入：nums = [0,1,1]
 * 输出：[]
 * 解释：唯一可能的三元组和不为 0 。
 * 示例 3：
 *
 * 输入：nums = [0,0,0]
 * 输出：[[0,0,0]]
 * 解释：唯一可能的三元组和为 0 。
 */
public class Solution0015 {

    public List<List<Integer>> threeSum(int[] nums) {
        Arrays.sort(nums);
        List<List<Integer>> result = new ArrayList<>();
        for(int k = 0; k < nums.length - 2; k ++) {
            if (nums[k] > 0) {
                break;
            }
            if (k > 0 && nums[k] == nums[k - 1]) {
                continue;
            }
            int i = k + 1, end = nums.length -1;
            while (i < end) {
                if (nums[k] + nums[i] + nums[end] == 0) {
                    result.add(Arrays.asList(nums[k], nums[i], nums[end]));
                    while(i < end && nums[end] == nums[--end]);
                    while(i < end && nums[i] == nums[++i]);
                } else if (nums[k] + nums[i] + nums[end] > 0) {
                    while(i < end && nums[end] == nums[--end]);
                } else {
                    while(i < end && nums[i] == nums[++i]);
                }
            }
        }
        return result;
    }

}
